requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
langchain
langchain-community
langchain-groq
langchain-google-genai
faiss-cpu
langchain-huggingface
sentence-transformers